Emission material. Speed up compilation and linking time for core, STL and vendor sources, using Precompiled-Header feature in CMake.

+++ include/api/memory_barrier.h
@@ -1,12 +1,10 @@
 #pragma once
 
-#include <core/common.h>
-
 #include <glad/glad.h>
 
 namespace gl {
 
-    enum MemoryBarrierBits : u8 {
+    enum MemoryBarrierBits : u32 {
         ALL = GL_ALL_BARRIER_BITS,
         SHADER_IMAGE_ACCESS = GL_SHADER_IMAGE_ACCESS_BARRIER_BIT
     };
